requirements.txt
SpeechRecognition==3.11.0
PyAudio==0.2.14
vosk==0.3.45
pyttsx3==2.98
PyYAML==6.0.2
numpy==2.1.3
scikit-learn==1.5.2